Extract ColumnMap
